Split calc and main
